esp32/PrimaryESP: Optimize & limit size of Pathfinder path list

+++ esp32/PrimaryESP32/alphabot/Pathfinder.h
@@ -23,8 +23,8 @@ struct PathNode {
 };
 
 struct Coordinate {
-    int32_t x;
-    int32_t y;
+    int16_t x;
+    int16_t y;
 };
 
 class Pathfinder {
